fix: fixed inconsistent vibration effects and other issues
